requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
lxml==4.9.3
python-dotenv==1.0.1
redis==5.0.1
markdown==3.5.1
